Split carousel and add it to exercise modal
